Fix transpose test gen of permutations & rank
Change transpose permutation generation to limit to the number of
possible permutations that can be created by the shape size or the
argument setting which ever is smaller. Also make sure all
permutations up to this number are generated rather than randomly
skipped due to duplicates.
Allow rank 1 transpose tests as the specification allows rank 1.

Change-Id: I28ea64c1d819f3af72c97bed43cfe7279c7e2f9c
Signed-off-by: Jeremy Johnson <jeremy.johnson@arm.com>