Add framework test for math.sign
The result comparison between Tensorflow runtime and the reference
model hasn't been checked as the sign operator is not supported by
the native TFLite runtime. That said, since the generated tosa ops
for tf.sign and tfl.sign is identical, the correctness presumably
can be proved by the result from tf.sign.

Change-Id: I72eb415df7fb6ca4dc9103f9ddc7104b0ba39234
Signed-off-by: TatWai Chong <tatwai.chong@arm.com>