Change version scheme to not guess next version
By default, setuptools_scm will guess the next version
and supply that as the version number when the package
is installed.

For example, this meant that installing
when tag "v0.60.2" was checked out would give package
version "0.60.3.dev8+gf5a8283.d20230920". With this
change, the package version would instead be
"0.60.2.post1.dev8+gad80c03".

Change-Id: I0bca9a6abec38e82839b52fe2933d2c0d14b5a84
Signed-off-by: Josef Malmström <josef.malmstrom@arm.com>