Shape infer dynamic model to static model prior to execution.
Dynamic shape model cannot directly run on the refenence model as
the concrete size of tensor is unknown therefore the volume of tensor
is not able to be allocated. Furthemore, the operators also expect
the input model is static-shaped.

This change turns dynamic model to static model prior to execution.

- Add `ifm_dynamic` field into json description to indicate whether
  the model has dynamic shape or not.

- Add the shape inference pass into the compilation pipeline, firstly
  legalize the dynamic tf/tfl model to dynamic tosa model with unknown
  shapes, and then run the shape inference pass with static shapes
  input argument to resolve unknown dimensions.

Change-Id: I5d2ffd452becc562dc30546789705bd01dd7a0b0
Signed-off-by: TatWai Chong <tatwai.chong@arm.com>