Update test generator to fix padding in depthwise_conv2d
Previously the padding was only fixed for conv2d. This was not creating
a problem with the used seed (42) but it already failed to create some
tests with other seeds and if the rng state was changed by adding code
using it. This is only a problem for ERROR_IF tests that often only
search for parameters in a very small parameter space.

Fixing up the padding for ERROR_IF tests made it so that we generated
too many tests for depthwise_conv2d. To counteract that, the selection
settings were changed to mimic the behaviour of conv2d --which doesn't
have this problem-- more closely.

Signed-off-by: IanTayler <ian.taylerlessa@arm.com>
Change-Id: I4ebf9538c3e3006d7172a1a70e8cb0f0b5e703cf