Extend SPECIAL data gen mode to Tensor ops
By evacha01 <evan.chandler@arm.com>:

Includes FP_SPECIAL mode for many ops, and adds a "broadcastable"
attribute to avoid fixing the shapes of operations that don't need it.

By iantay01 <ian.taylerlessa@arm.com>:

Fixes the comparison of special values to ignore sign of 0 and avoids
MIN/MAX float values for operations in which they overflow.

Changes a warning that was raised when no `_s0` test existed
to instead only warn if _s0` doesn't exist but some `_sN` does. This
allows for tests that would not make sense for dot products due to a
small sample size in the statistical analysis, but could still be done
for fp_special where no statistical analysis is performed.

Changes the padding value in OpMaxPool2d from `-max` to `-inf` in
floating point types. In the future, a more thorough check of other
uses of `-max` and `+max` in place of `-inf` and `+inf` should be
performed.

Implements a `DtypeLimits` template struct similar to
`std::numeric_limits` to support the proper limits for the floating
point types we have as dtypes but have not implemented as a C++ type
in the reference model yet.

Re-implements max_pool2d and argmax by hand to circumvent bug in Eigen's
argmax that was failing to recognize `-max` as the maximum value in an
axis with otherwise only `-inf`.

Co-authored-by: evacha01 <evan.chandler@arm.com>
Signed-off-by: IanTayler <ian.taylerlessa@arm.com>
Change-Id: I429aea90a962d9ff2fd2eb45cc9d6c8f5bf702ec