Fix bounds in absolute error verification unit tests
The "positive - abs error" test failed in some RNG states. An example
was when changing the seed of the generator to 10 and running all tests
This was because the code did not account for the floating point error
of the operations it was performing to find the test values.

That was fixed by both reducing the error that is introduced and
accounting for it.

Signed-off-by: Ian Tayler Lessa <ian.taylerlessa@arm.com>
Change-Id: Ic758be76ef7204633cc7ead84b346879d3f0c459