Subnormal support changes for FP8, FP32, FP16 and BF16.
Update the FP bounds check to require subnormals for FP8 types
and only allow results to be flushed to zero for non-FP8.

Introduce support for subnormals to be flushed to zero
for FP32, FP16 and BF16 types for "exact" operations like
MINIMUM.
Implemented by re-using the existing bounds check but with bounds
of zero to allow for subnormals to be flushed to zero by an
implementation.

Change-Id: Ic0d8e74252d37212633290d9f39aa6fc74a327f3
Signed-off-by: Jeremy Johnson <jeremy.johnson@arm.com>