Ignore casting warning in generate_fp_special.cc
A clang warning about a loss of precision in a cast can be safely
ignored for our purposes when generating FP_SPECIAL values as all we are
looking for is a value that is higher than the maximum INT32::MAX value,
and we will get it with or without the loss of precision in the implicit
cast to float.

Added TODO for future improvements to cfloat.h allowing casting from
other types including int32_t.

Signed-off-by: Ian Tayler Lessa <ian.taylerlessa@arm.com>
Change-Id: Iebfb47fdc3e2d26e5e27fe7b966431804dfc925e