Conditionally Build Verfy Utility Tool
The `cxxopts` library is a dependency of `tosa_verify` which is only
fetched when `BUILD_TOSA_REFERENCE_MODEL_EXECUTABLE` is `ON` hence we
only build this target when the same condition is satisfied.

Fix some CMake hygeine issues.

Add a regex to the .gitignore file which includes build directories of
the form `build-debug`, `build-release`, `build-asan` etc.

Signed-off-by: Jack Frankland <jack.frankland@arm.com>
Change-Id: I10dc79435cf2fe9c7ad3da17e31f0d133ff714f1