Fix qinfo usage in CONV ops and MATMUL
We were regenerating qinfo in the tvg functions without respecting the
CLI arguments for the builder, and in the case of MATMUL we were using
the wrong values to verify ERROR_IF tests.
This patch fixes those issues by using the same qinfo in tvg functions
as we use everywhere else.

Signed-off-by: Ian Tayler Lessa <ian.taylerlessa@arm.com>
Change-Id: I3a7412bd83ceb40f0c8bbc6adfd824069baaddb1