Fix allocation of empty shapes
Previously we were allocating them as rank 0 tensors, but that requires
us to make some meaningless data associated with it in the flatbuffer.

Now we accept the representation `shape: [], data: []` for the empty
shape [] (i.e. the shape of a rank 0 tensor).

Now we also read the shape input in RESHAPE instead of using the shape
of the output tensor.

Add test_dynamic_linear_14x19_f32 to the ignore list for framework tests
as it includes a reshape to a dynamic shape.

Signed-off-by: Ian Tayler Lessa <ian.taylerlessa@arm.com>
Change-Id: Iab93590cdd30744853477d97b67acdececeda799