Check multipliers and shift sizes in RESCALE
OpRescale::checkTensorAttributes was missing checks related to the
relationship between the size of multiplier and shift arrays and the
value of the per_channel attribute.

This patch adds those checks and adds some tests to catch regressions in
the future.

Some cleanup of the ERROR_IF code was also done with no expected
change in behaviour.

Signed-off-by: Ian Tayler Lessa <ian.taylerlessa@arm.com>
Change-Id: I0b7a071af61c84fb733600a9f71f5ab7b1d754fc