Avoid supplying unsupported linker flags
This commit aims to fix the MacOS build of the
reference model by, more generally, ensuring
unsupported linker flags are not supplied to the
selected linker.

Note: The "check_linker_flag" utility is used by
this change which requires cmake >= 3.18,
therefore, the minimum supported cmake version in
this project has been bumped from 3.16 to 3.18.

Change-Id: I17b1c176d328088554c231539ddfe3b057914547
Signed-off-by: Luke Hutton <luke.hutton@arm.com>